Move view tests out of controllers
